django_app/requirements.txt
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-extensions==3.2.3
django-phonenumber-field==7.1.0
django-split-settings==1.2.0
idna==3.4
kombu==5.3.2
mysqlclient==2.2.0
phonenumbers==8.13.19
prompt-toolkit==3.0.39
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uWSGI==2.0.22
vine==5.0.0
wcwidth==0.2.6
